[console] prototype of a silex console
currently there is only one single `update` command:

    php silex.phar update

this will fetch the latest version of the phar from silex-project.org
and replace the current one with it.

this makes it easier for users because they can easily update. we could
potentially even provide a changelog on running the command.

this commit also introduces a Silex\Application::VERSION constant that
currently is set to the current git SHA-1, as I proposed in PR 33. This
is done by the Compiler.

unfortunately it does increase the size of the PHAR from 387K to 467K
(because of inclusion of Console and Process components). Creating a
light-weight version without those components is certainly an option, but
not needed imo.