[console] lightweight silex console
you can get the current version by calling the phar without arguments:

    $ php silex.phar
    Silex version 1ee6b36

there is only one single `update` command:

    $ php silex.phar update

this will fetch the latest version of the phar from silex-project.org
and replace the current one with it.

this makes it easier for users because they can easily update. we could
potentially even provide a changelog on running the command.

this commit also introduces a Silex\Application::VERSION constant that
currently is set to the current git SHA-1, as I proposed in PR 33. This
is done by the Compiler. It can later be adjusted to use `git describe`.

this is an implementation of PR 35 without the console component.