Support including a phar from another phar
The default stubs mess up when you try to run a phar which includes
another phar. In my case I was running pip.phar on the CLI, which
included app.php, which in turn included silex.phar.

It broke the __DIR__ calls. This was able to fix it.

Needs testing in lots of different setups, because the default stubs are
no longer used.