moved the setting of provider values after the registration to avoid having isset calls everywhere in the provider code
Before, the behavior was not consistent. Some values could be overridden
by passing an array as a second argument to the register() method, but
for most of them it did not work (we would have to wrap all definitions
with an isset()). Now, this is more consistent as you can override
everything that is defined in the provider.