Fix PHP 5.4 built-in webserver front controller
The problem is that with `php -S localhost:8080 -t web` it will only route
one level of paths to index.php. This is stupid, but requires us to use
a router script, which means that we must specify `index.php` and have
that script return false when the web server should serve from disk.

Funny part: We are almost lying now, because the webserver now requires
some sort of configuration (via code), but I will leave it in the docs
for marketing reasons. It's not a big lie after all.