Add minimum-stability flag to all composer.json samples
Today marks the release of composer alpha4. This changes the default
stability from dev to stable, which means that if you want to use un-
stable dependencies, you either have to specify a root package level
minimum-stability, or use the @dev constraint, i.e. "1.0.*@dev".

We are setting the minimum-stability, so that we do not have to use
@dev for all dependencies, and all dependencies of dependencies.

We can increase the stability to beta once we release a beta of silex
(this will use Symfony 2.1 beta releases), and when we make a final
release (some time after Symfony 2.1.0), the flag can be removed.