Patch Application.php to fix before filters.
It looks like there are some strange things going on with event handler
priorities in the Symfony code.  From what I can tell, there is an interaction
between the RouterListener and part of the Secuirty Component that requires
specific priorities on their respective handler methods.  Unfortunately, this
seems to break some assumptions in the way that Silex is registering its own
handler method for before filters.