refactored Silex events (refs #503, refs #504, refs #452)
Removed the SilexEvents class. Using the standard Symfony events allows
us to be more flexible (the priorities are the same as the one from
Symfony, so you can hook at any place into the HTTP handling).

Updated documentation to make it accurate with the current behavior, and
extended it to explain the flexibility the developer has.