moved all subscribers registration into the dispatcher closure creation
This removes yet another special case that should not exist. Now, all
listeners/subscribers registration are done in one place.

If you want to disable the exception_handler, you now should call the
disable() method on it. You can still unset it but be careful to do it
as early as possible (which is anyway always the best idea).