note that `mod_rewrite` needs to be enabled
If `mod_rewrite` isn't enabled - like in default Apache installs and WAMP - it will result in 404's. I had this issue with the hello world zip distribution (SQ question: http://stackoverflow.com/questions/15498419/why-am-i-getting-these-silex-404s) and it was quite difficult to track down. Almost made me give up on Silex, although this is of course not a Silex issue - but it can be challenging for beginners, and if you can't get the hello world to run it kinda takes the wind out of your sails.