Allow strings for levels in the monolog service provider
The monolog service provider requires a constant on Monolog\Logger to be
referenced, which has an integer value. This is not very practical when
working with external configuration files.

This patch accepts strings like debug, info, error, etc. that are
translated to the corresponding logger constant values.

This is an alternative to #882.