Fixes #1017 by cascading __call to mounted ControllerCollection
all method calls that are handled by ControllerCollection::__call
will be recorded and replayed for a mounted ControllerCollection.

forward method calls to already mounted ControllerCollections
in the list of controllers.

include support for route-level middlewares using a priority
queue for the callbacks. insertion order is preserved by converting
the priority to an array and negative count as the subordinate
priority. the priority queue is cloned before being iterated in the
middleware listener to allow multiple iterations (iterating the queue
removes its items)