Allow custom logout handlers when using custom authentication
When using a custom authentication factory:
As of now, one can only select a Logout Success Handler(which is erroneously called logout handler, but I left the old name for compatibility's sake) which is called after the Logout handler was called and the logout was successful.
But custom authentication may need a custom Logout handler.(one may not want to use sessions at all, but custom cookies/etc.)
I've called it logout_clear_data_handler, since it clears data. And by default it should be SessionLogoutHandler();(the prototype for it).