Removed overwriting of $app['swiftmailer.options']
When registering this provider, $app['swiftmailer.options'] is reset to an empty array so any configuration has to be loaded after the provider has been registered (which is not the same behavior as the other providers). In line 64, we are expecting the same array to be populated and replaced into the final options array.